requirements.txt
mcp>=0.1.1
graphql-core>=3.2.3
numpy>=1.26.4
python-dotenv>=1.0.1
aiohttp>=3.9.5
